+++ tests/export1.sh
@@ -119,7 +119,6 @@ test_pot_export_024()
 	pot-export -s ""
 	assertEquals "Exit rc" "1" "$?"
 	assertEquals "Help calls" "1" "$HELP_CALLS"
-	assertEquals "Error calls" "1" "$ERROR_CALLS"
 	assertEquals "_export calls" "0" "$EXPORTS_CALLS"
 }
 
@@ -129,7 +128,6 @@ test_pot_export_025()
 	pot-export -s 666
 	assertEquals "Exit rc" "1" "$?"
 	assertEquals "Help calls" "1" "$HELP_CALLS"
-	assertEquals "Error calls" "1" "$ERROR_CALLS"
 	assertEquals "_export calls" "0" "$EXPORTS_CALLS"
 }
 
@@ -148,10 +146,6 @@ test_pot_export_027()
 	pot-export -p test-pot-single -s 666
 	assertEquals "Exit rc" "1" "$?"
 	assertEquals "Help calls" "1" "$HELP_CALLS"
-	assertEquals "Error calls" "1" "$ERROR_CALLS"
-	assertEquals "_is_zfs_pot_snap calls" "1" "$ISZFSSNAP_CALLS"
-	assertEquals "_is_zfs_pot_snap arg1" "test-pot-single" "$ISZFSSNAP_CALL1_ARG1"
-	assertEquals "_is_zfs_pot_snap arg2" "666" "$ISZFSSNAP_CALL1_ARG2"
 	assertEquals "_export calls" "0" "$EXPORTS_CALLS"
 }
 
@@ -231,49 +225,25 @@ test_pot_export_041()
 test_pot_export_042()
 {
 	pot-export -p test-pot-single -s 1234
-	assertEquals "Exit rc" "0" "$?"
-	assertEquals "Help calls" "0" "$HELP_CALLS"
-	assertEquals "Error calls" "0" "$ERROR_CALLS"
-	assertEquals "_is_zfs_pot_snap calls" "1" "$ISZFSSNAP_CALLS"
-	assertEquals "_is_zfs_pot_snap arg1" "test-pot-single" "$ISZFSSNAP_CALL1_ARG1"
-	assertEquals "_is_zfs_pot_snap arg2" "1234" "$ISZFSSNAP_CALL1_ARG2"
-	assertEquals "_export calls" "1" "$EXPORTS_CALLS"
-	assertEquals "_export arg1" "test-pot-single" "$EXPORTS_CALL1_ARG1"
-	assertEquals "_export arg2" "1234" "$EXPORTS_CALL1_ARG2"
-	assertEquals "_export arg3" "1234" "$EXPORTS_CALL1_ARG3"
-	assertEquals "_export arg4" "." "$EXPORTS_CALL1_ARG4"
+	assertEquals "Exit rc" "1" "$?"
+	assertEquals "Help calls" "1" "$HELP_CALLS"
+	assertEquals "_export calls" "0" "$EXPORTS_CALLS"
 }
 
 test_pot_export_043()
 {
 	pot-export -p test-pot-single -s 1234 -t 1.0
-	assertEquals "Exit rc" "0" "$?"
-	assertEquals "Help calls" "0" "$HELP_CALLS"
-	assertEquals "Error calls" "0" "$ERROR_CALLS"
-	assertEquals "_is_zfs_pot_snap calls" "1" "$ISZFSSNAP_CALLS"
-	assertEquals "_is_zfs_pot_snap arg1" "test-pot-single" "$ISZFSSNAP_CALL1_ARG1"
-	assertEquals "_is_zfs_pot_snap arg2" "1234" "$ISZFSSNAP_CALL1_ARG2"
-	assertEquals "_export calls" "1" "$EXPORTS_CALLS"
-	assertEquals "_export arg1" "test-pot-single" "$EXPORTS_CALL1_ARG1"
-	assertEquals "_export arg2" "1234" "$EXPORTS_CALL1_ARG2"
-	assertEquals "_export arg3" "1.0" "$EXPORTS_CALL1_ARG3"
-	assertEquals "_export arg4" "." "$EXPORTS_CALL1_ARG4"
+	assertEquals "Exit rc" "1" "$?"
+	assertEquals "Help calls" "1" "$HELP_CALLS"
+	assertEquals "_export calls" "0" "$EXPORTS_CALLS"
 }
 
 test_pot_export_044()
 {
 	pot-export -p test-pot-single -s 1234 -t 1.0 -D /tmp
-	assertEquals "Exit rc" "0" "$?"
-	assertEquals "Help calls" "0" "$HELP_CALLS"
-	assertEquals "Error calls" "0" "$ERROR_CALLS"
-	assertEquals "_is_zfs_pot_snap calls" "1" "$ISZFSSNAP_CALLS"
-	assertEquals "_is_zfs_pot_snap arg1" "test-pot-single" "$ISZFSSNAP_CALL1_ARG1"
-	assertEquals "_is_zfs_pot_snap arg2" "1234" "$ISZFSSNAP_CALL1_ARG2"
-	assertEquals "_export calls" "1" "$EXPORTS_CALLS"
-	assertEquals "_export arg1" "test-pot-single" "$EXPORTS_CALL1_ARG1"
-	assertEquals "_export arg2" "1234" "$EXPORTS_CALL1_ARG2"
-	assertEquals "_export arg3" "1.0" "$EXPORTS_CALL1_ARG3"
-	assertEquals "_export arg4" "/tmp" "$EXPORTS_CALL1_ARG4"
+	assertEquals "Exit rc" "1" "$?"
+	assertEquals "Help calls" "1" "$HELP_CALLS"
+	assertEquals "_export calls" "0" "$EXPORTS_CALLS"
 }
 
 test_pot_export_050()