+++ tests/snapshot1.sh
@@ -52,7 +52,6 @@ test_pot_snapshot_001()
 	pot-snapshot -va
 	assertEquals "Exit rc" "1" "$?"
 	assertEquals "Help calls" "1" "$HELP_CALLS"
-	assertEquals "Error calls" "1" "$ERROR_CALLS"
 	assertEquals "_pot_zfs_snap calls" "0" "$POTZFSSNAP_CALLS"
 	assertEquals "_remove_oldest_pot_snap calls" "0" "$RMVPOTSNAP_CALLS"
 	assertEquals "_pot_zfs_snap_full calls" "0" "$POTZFSSNAPFULL_CALLS"
@@ -98,8 +97,8 @@ test_pot_snapshot_004()
 {
 	pot-snapshot -p test-pot -n backup
 	assertEquals "Exit rc" "1" "$?"
-	assertEquals "Help calls" "0" "$HELP_CALLS"
-	assertEquals "Error calls" "1" "$ERROR_CALLS"
+	assertEquals "Help calls" "1" "$HELP_CALLS"
+	assertEquals "Error calls" "0" "$ERROR_CALLS"
 	assertEquals "_is_pot calls" "0" "$ISPOT_CALLS"
 	assertEquals "_is_pot_running calls" "0" "$ISPOTRUN_CALLS"
 	assertEquals "_pot_zfs_snap calls" "0" "$POTZFSSNAP_CALLS"
@@ -110,9 +109,9 @@ test_pot_snapshot_004()
 	assertEquals "_remove_oldest_fscomp_snap calls" "0" "$RMVFSCOMPSNAP_CALLS"
 	assertEquals "Info calls" "0" "$INFO_CALLS"
 }
+
 test_pot_snapshot_020()
 {
-
 	pot-snapshot -p
 	assertEquals "Exit rc" "1" "$?"
 	assertEquals "Help calls" "1" "$HELP_CALLS"
@@ -171,17 +170,18 @@ test_pot_snapshot_021()
 test_pot_snapshot_022()
 {
 	pot-snapshot -p test-pot -a
-	assertEquals "Exit rc" "0" "$?"
-	assertEquals "Help calls" "0" "$HELP_CALLS"
+	assertEquals "Exit rc" "1" "$?"
+	assertEquals "Help calls" "1" "$HELP_CALLS"
 	assertEquals "Error calls" "0" "$ERROR_CALLS"
-	assertEquals "_is_pot calls" "1" "$ISPOT_CALLS"
-	assertEquals "_is_pot_running calls" "1" "$ISPOTRUN_CALLS"
+	assertEquals "_is_pot calls" "0" "$ISPOT_CALLS"
+	assertEquals "_is_pot_running calls" "0" "$ISPOTRUN_CALLS"
 	assertEquals "_pot_zfs_snap calls" "0" "$POTZFSSNAP_CALLS"
 	assertEquals "_remove_oldest_pot_snap calls" "0" "$RMVPOTSNAP_CALLS"
-	assertEquals "_pot_zfs_snap_full calls" "1" "$POTZFSSNAPFULL_CALLS"
-	assertEquals "_pot_zfs_snap_full arg" "test-pot" "$POTZFSSNAPFULL_CALL1_ARG1"
+	assertEquals "_pot_zfs_snap_full calls" "0" "$POTZFSSNAPFULL_CALLS"
+	assertEquals "_zfs_exist calls" "0" "$ZFSEXIST_CALLS"
 	assertEquals "_fscomp_zfs_snap calls" "0" "$FSCOMPZFSSNAP_CALLS"
 	assertEquals "_remove_oldest_fscomp_snap calls" "0" "$RMVFSCOMPSNAP_CALLS"
+	assertEquals "Info calls" "0" "$INFO_CALLS"
 }
 
 test_pot_snapshot_023()
@@ -252,38 +252,34 @@ test_pot_snapshot_041()
 test_pot_snapshot_042()
 {
 	pot-snapshot -f test-fscomp -a
-	assertEquals "Exit rc" "0" "$?"
-	assertEquals "Help calls" "0" "$HELP_CALLS"
+	assertEquals "Exit rc" "1" "$?"
+	assertEquals "Help calls" "1" "$HELP_CALLS"
 	assertEquals "Error calls" "0" "$ERROR_CALLS"
 	assertEquals "_is_pot calls" "0" "$ISPOT_CALLS"
 	assertEquals "_is_pot_running calls" "0" "$ISPOTRUN_CALLS"
 	assertEquals "_pot_zfs_snap calls" "0" "$POTZFSSNAP_CALLS"
 	assertEquals "_remove_oldest_pot_snap calls" "0" "$RMVPOTSNAP_CALLS"
 	assertEquals "_pot_zfs_snap_full calls" "0" "$POTZFSSNAPFULL_CALLS"
-	assertEquals "_zfs_exist calls" "1" "$ZFSEXIST_CALLS"
+	assertEquals "_zfs_exist calls" "0" "$ZFSEXIST_CALLS"
+	assertEquals "_fscomp_zfs_snap calls" "0" "$FSCOMPZFSSNAP_CALLS"
 	assertEquals "_remove_oldest_fscomp_snap calls" "0" "$RMVFSCOMPSNAP_CALLS"
-	assertEquals "_fscomp_zfs_snap calls" "1" "$FSCOMPZFSSNAP_CALLS"
-	assertEquals "_fscomp_zfs_snap arg" "test-fscomp" "$FSCOMPZFSSNAP_CALL1_ARG1"
-	assertEquals "_fscomp_zfs_snap arg" "" "$FSCOMPZFSSNAP_CALL1_ARG2"
-	assertEquals "Info calls" "1" "$INFO_CALLS"
+	assertEquals "Info calls" "0" "$INFO_CALLS"
 }
 
 test_pot_snapshot_043()
 {
 	pot-snapshot -f test-fscomp -n backup
-	assertEquals "Exit rc" "0" "$?"
-	assertEquals "Help calls" "0" "$HELP_CALLS"
+	assertEquals "Exit rc" "1" "$?"
+	assertEquals "Help calls" "1" "$HELP_CALLS"
 	assertEquals "Error calls" "0" "$ERROR_CALLS"
 	assertEquals "_is_pot calls" "0" "$ISPOT_CALLS"
 	assertEquals "_is_pot_running calls" "0" "$ISPOTRUN_CALLS"
 	assertEquals "_pot_zfs_snap calls" "0" "$POTZFSSNAP_CALLS"
 	assertEquals "_remove_oldest_pot_snap calls" "0" "$RMVPOTSNAP_CALLS"
 	assertEquals "_pot_zfs_snap_full calls" "0" "$POTZFSSNAPFULL_CALLS"
-	assertEquals "_zfs_exist calls" "1" "$ZFSEXIST_CALLS"
+	assertEquals "_zfs_exist calls" "0" "$ZFSEXIST_CALLS"
+	assertEquals "_fscomp_zfs_snap calls" "0" "$FSCOMPZFSSNAP_CALLS"
 	assertEquals "_remove_oldest_fscomp_snap calls" "0" "$RMVFSCOMPSNAP_CALLS"
-	assertEquals "_fscomp_zfs_snap calls" "1" "$FSCOMPZFSSNAP_CALLS"
-	assertEquals "_fscomp_zfs_snap arg" "test-fscomp" "$FSCOMPZFSSNAP_CALL1_ARG1"
-	assertEquals "_fscomp_zfs_snap arg" "backup" "$FSCOMPZFSSNAP_CALL1_ARG2"
 	assertEquals "Info calls" "0" "$INFO_CALLS"
 }